TRegex: loosen max parser tree size.

+++ dataset/GitHub_Java/oracle.graal/regex/src/com.oracle.truffle.regex/src/com/oracle/truffle/regex/tregex/automaton/SmallStateSetImpl.java
@@ -46,6 +46,10 @@ import java.util.Set;
 
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 
+/**
+ * Specialized implementation of {@link StateSet} for {@link StateIndex state indices} with less
+ * than 65 elements. Uses a single {@code long} value as a bit set.
+ */
 final class SmallStateSetImpl<S> implements StateSet<S> {
 
     private final StateIndex<? super S> stateIndex;
@@ -86,7 +90,7 @@ final class SmallStateSetImpl<S> implements StateSet<S> {
         return toBit(stateIndex.getId((S) o));
     }
 
-    private static long toBit(short id) {
+    private static long toBit(int id) {
         assert 0 <= id && id < 64;
         return 1L << id;
     }
@@ -104,7 +108,7 @@ final class SmallStateSetImpl<S> implements StateSet<S> {
         return notPresent;
     }
 
-    private void removeId(short id) {
+    private void removeId(int id) {
         assert (set & toBit(id)) != 0;
         set &= ~toBit(id);
     }
@@ -218,6 +222,9 @@ final class SmallStateSetImpl<S> implements StateSet<S> {
         return defaultToString();
     }
 
+    /**
+     * Yields all elements in this set, ordered by their {@link StateIndex#getId(Object) ID}.
+     */
     @Override
     public Iterator<S> iterator() {
         return new SmallStateSetIterator<>(this, set);
@@ -253,7 +260,7 @@ final class SmallStateSetImpl<S> implements StateSet<S> {
         @Override
         public void remove() {
             assert bitIndex > 0;
-            stateSet.removeId((short) (bitIndex - 1));
+            stateSet.removeId(bitIndex - 1);
         }
     }
 }
